requirements.txt
Keras==2.0.5
numpy==1.13.3
tweet-preprocessor==0.5.0
nltk==3.2.2
autocorrect==0.3.0
tensorflow==1.0.1
Flask==0.12
flask-cors==3.0.3
h5py==2.7.1